+++ contracts/token/ERC20/extensions/IERC20Votes.sol
@@ -6,7 +6,7 @@ import "../IERC20.sol";
 
 interface IERC20Votes is IERC20 {
     struct Checkpoint {
-        uint32  fromBlock;
+        uint32 fromBlock;
         uint224 votes;
     }
 
@@ -14,11 +14,25 @@ interface IERC20Votes is IERC20 {
     event DelegateVotesChanged(address indexed delegate, uint256 previousBalance, uint256 newBalance);
 
     function delegates(address owner) external view returns (address);
+
     function checkpoints(address account, uint32 pos) external view returns (Checkpoint memory);
+
     function numCheckpoints(address account) external view returns (uint32);
+
     function getCurrentVotes(address account) external view returns (uint256);
+
     function getPriorVotes(address account, uint256 blockNumber) external view returns (uint256);
-    function getPriorTotalSupply(uint256 blockNumber) external view returns(uint256);
+
+    function getPriorTotalSupply(uint256 blockNumber) external view returns (uint256);
+
     function delegate(address delegatee) external;
-    function delegateBySig(address delegatee, uint nonce, uint expiry, uint8 v, bytes32 r, bytes32 s) external;
+
+    function delegateBySig(
+        address delegatee,
+        uint256 nonce,
+        uint256 expiry,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external;
 }
